Local movie database roughed in. It persists movies marked as favorite.  Passes basic unit tests. If the user later un-favorites the movie, it remains in the database, but is no longer marked a favorite.  I'm doing this because chances are the user / tester is going to poke the favorite star multiple times for the same movie.  Might as well have it perform well.  Over time there may be more movies in the database then needed, so I'll need to put a clean up command in the settings. Speaking of cleanup - there is some rigamarole with mDatabaseExists that is probably not necessary. Minor UI change: set semi-opaque background behind title and favorite star in detail view so the star shows up better. Next up - persist the favorite movie posters on the device.

+++ PopularMovies/mobile/src/main/java/com/android/bazemom/popularmovies/moviemodel/MovieDetailModel.java
@@ -39,7 +39,7 @@ public class MovieDetailModel {
     @Expose
     private String overview;
     @Expose
-    private double popularity;
+    private float popularity;
     @SerializedName("poster_path")
     @Expose
     private String posterPath;
@@ -69,7 +69,7 @@ public class MovieDetailModel {
     private boolean video;
     @SerializedName("vote_average")
     @Expose
-    private double voteAverage;
+    private float voteAverage;
     @SerializedName("vote_count")
     @Expose
     private int voteCount;
@@ -286,7 +286,7 @@ public class MovieDetailModel {
      * @param popularity
      * The popularity
      */
-    public void setPopularity(double popularity) {
+    public void setPopularity(float popularity) {
         this.popularity = popularity;
     }
 
@@ -502,7 +502,7 @@ public class MovieDetailModel {
      * @param voteAverage
      * The vote_average
      */
-    public void setVoteAverage(double voteAverage) {
+    public void setVoteAverage(float voteAverage) {
         this.voteAverage = voteAverage;
     }
